requirements.txt
opencv-python
numpy
bindglobal
pillow